Users can only view their own lists

+++ mysite/main/views.py
@@ -11,37 +11,41 @@ def index(response, id):
 
     ls = ToDoList.objects.get(id=id)
 
-    #{"save":["save"], "c1":["clicked"]}
-    if response.method == "POST":
-        print(response.POST)
-        print("\n")
+    if ls in response.user.todolist.all():
+        if response.method == "POST":
+            print(response.POST)
+            print("\n")
 
-        # Delete selected items
-        if response.POST.get("delete"):
-            print("Delete method")
-            for item in ls.item_set.all():
-                if response.POST.get("c" + str(item.id)) == "clicked":
-                    item.delete()
+            # Delete selected items
+            if response.POST.get("delete"):
+                print("Delete method")
+                for item in ls.item_set.all():
+                    if response.POST.get("c" + str(item.id)) == "clicked":
+                        item.delete()
 
-        # Save changes to items
-        if response.POST.get("save"):
-            for item in ls.item_set.all():
-                if response.POST.get("c" + str(item.id)) == "clicked":
-                    item.complete = True
-                else: 
-                    item.complete = False
-                item.save()
-        
-        # Add new item 
-        elif response.POST.get("newItem"):
-            txt = response.POST.get("new")
-            print(len(txt))
-            if len(txt) > 2:
-                ls.item_set.create(text=txt, complete=False)
-            else:
-                print("invalid input")
-        
-    return render(response, "main/list.html", {"ls":ls})
+            # Save changes to items
+            if response.POST.get("save"):
+                for item in ls.item_set.all():
+                    if response.POST.get("c" + str(item.id)) == "clicked":
+                        item.complete = True
+                    else: 
+                        item.complete = False
+                    item.save()
+            
+            # Add new item 
+            elif response.POST.get("newItem"):
+                txt = response.POST.get("new")
+                print(len(txt))
+                if len(txt) > 2:
+                    ls.item_set.create(text=txt, complete=False)
+                else:
+                    print("invalid input")
+            
+        return render(response, "main/list.html", {"ls":ls})
+
+    else:
+        return HttpResponseRedirect("/view")
+
 
 def home(response):
     print("HOME")
